Refactor manager/machine.go -> utils/machine/info.go
